src/requirements.txt
folium==0.16.0
geopandas==0.14.4
mysql-connector-python
boto3

